+++ docs/changelog.rst
@@ -8,6 +8,8 @@ Changelog
 
 - Removed several features in favor of keeping the API simpler
   and the code more maintainable
+- Stricter one-to-one checking checking by default
+- New :attr:`bidict.bidict.forceupdate`` method for bulk forceput
 
 Breaking API Changes
 ^^^^^^^^^^^^^^^^^^^^
@@ -21,6 +23,10 @@ Breaking API Changes
 - Removed ``bidict.invert``.
   Use ``b.inv`` rather than inverting a bidict in place.
   `#20 <https://github.com/jab/bidict/issues/20>`_
+- Raise :class:`bidict.ValueExistsException`` when attempting to insert a new
+  key associated with an existing value
+- Rename ``collapsingbidict`` to :class:`bidict.loosebidict`` now that it's
+  loose in the case of :class:`bidict.ValueExistsException`` too
 
 
 0.9.0.post1 (2015-06-06)